\name{BiocFoundationGPU-package}
\alias{BiocFoundationGPU-package}
\alias{BiocFoundationGPU}
\docType{package}
\title{
\packageTitle{BiocFoundationGPU}
}
\description{
\packageDescription{BiocFoundationGPU}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{BiocFoundationGPU}

Maintainer: \packageMaintainer{BiocFoundationGPU}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = BiocFoundationGPU)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{BiocFoundationGPU}}
%\section{Documentation Index}{\packageIndices{BiocFoundationGPU}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
