% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{Run_scGPT}
\alias{Run_scGPT}
\title{scGPT}
\usage{
Run_scGPT(h5ad_file, model_dir, gene_col, batch_size = 64L)
}
\arguments{
\item{h5ad_file}{file that points to the location of a h5ad file with the
count data}

\item{model_dir}{directory with the model. This should be a directory that
the scGPT model, consisting of a `.pt` file with the model, as well as the
`vocab.json` and `args.json` file.}

\item{gene_col}{The gene column name}

\item{batch_size}{The batch size passed to `scg.tasks.embed_data` (default 64)}
}
\description{
This function runs scGPT.
}
\details{
This function runs the `scg.tasks.embed_data` function
}
